Rename EditingState as SettingsState
