+++ client/components/admin/settings/groups/GenericGroupPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { useGroup } from '../EditingState';
+import { useGroup } from '../SettingsState';
 import { GroupPage } from '../GroupPage';
 import { Section } from '../Section';
